Moved MDAnalysis.KDTree to MDAnalysis.lib.KDTree
Part of Issue #287
