+++ package/MDAnalysis/KDTree.py
@@ -1,0 +1,9 @@
+# Placeholder because KDTree moved
+__all__ = ['KDTree', 'NeighborSearch']
+
+import warnings
+warnings.simplefilter('always', DeprecationWarning)
+warnings.warn("KDTree has moved to MDAnalysis.lib.KDTree",
+              DeprecationWarning)
+
+from .lib.KDTree import KDTree, NeighborSearch
