+++ package/doc/sphinx/source/documentation_pages/lib_modules.rst
@@ -12,4 +12,5 @@ Versions of these routines with error checking are available in
 .. toctree::
    :maxdepth: 1
 
-   lib/_distances
+   ./lib/KDTree_modules
+   ./lib/_distances
