+++ package/setup.py
@@ -245,7 +245,7 @@ if __name__ == '__main__':
                   define_macros=define_macros,
                   include_dirs=include_dirs,
                   extra_compile_args=extra_compile_args),
-        Extension('KDTree._CKDTree',
+        Extension('lib.KDTree._CKDTree',
                   ["src/KDTree/KDTree.cpp",
                       "src/KDTree/KDTree.swig.cpp"],
                   include_dirs=include_dirs,
@@ -282,8 +282,8 @@ if __name__ == '__main__':
                     'MDAnalysis.coordinates.pdb',
                     'MDAnalysis.core',
                     'MDAnalysis.core.parallel',
-                    'MDAnalysis.KDTree',
                     'MDAnalysis.lib',
+                    'MDAnalysis.lib.KDTree',
                     'MDAnalysis.selections',
                     'MDAnalysis.topology',
                     'MDAnalysis.topology.tpr',
